backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
SQLAlchemy==2.0.32
sqlparse==0.5.1
python-dotenv==1.0.1
jinja2==3.1.4
httpx==0.27.0
pyyaml==6.0.1
pandas==2.2.2

streamlit==1.37.0
